requirements.txt
Django==6.0

# Production WSGI server
gunicorn==22.0.0

# Static files in production
whitenoise==6.7.0

# Database URL parsing (Render Postgres)
dj-database-url==2.2.0

# PostgreSQL driver (psycopg3)
psycopg[binary]==3.2.4

# Image processing
pillow==12.0.0

# PDF generation and parsing
reportlab==4.4.7
pypdf==6.5.0

# HTML parsing (SEO utilities)
beautifulsoup4==4.14.3